requirements.txt
swarmnode==0.1.0
fastapi==0.104.1
uvicorn==0.23.2
langchain==0.0.335
langchain-openai==0.0.2
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.4.2 